Moving mod/babel to src/Module/babel
